feat(linter): valid_expect_in_promise vitest rule
