+++ crates/oxc_linter/src/config/config_store.rs
@@ -904,7 +904,7 @@ mod test {
             {
                 "plugins": ["vitest", "typescript"],
                 "rules": {
-                    "vitest/valid-expect-in-promise": "error",
+                    "vitest/no-test-prefixes": "error",
                     "typescript/no-explicit-any": "error"
                 },
                 "overrides": [
@@ -920,8 +920,10 @@ mod test {
         let rules_for_test_file = config.apply_overrides("foo.test.ts".as_ref());
 
         assert!(
-            rules_for_test_file.rules.iter().any(|(rule, _)| rule.plugin_name() == "jest"
-                && rule.name() == "valid-expect-in-promise"),
+            rules_for_test_file
+                .rules
+                .iter()
+                .any(|(rule, _)| rule.plugin_name() == "jest" && rule.name() == "no-test-prefixes"),
             "vitest-compatible jest rules should remain enabled when an override matches"
         );
         assert!(
