+++ crates/oxc_linter/src/rules.rs
@@ -765,6 +765,7 @@ pub(crate) mod vitest {
     pub mod require_test_timeout;
     pub mod require_top_level_describe;
     pub mod valid_expect;
+    pub mod valid_expect_in_promise;
     pub mod valid_title;
     pub mod warn_todo;
 }
