+++ crates/oxc_linter/src/utils/mod.rs
@@ -37,7 +37,7 @@ pub use self::{
 // the crates/oxc_linter/data/vitest_compatible_jest_rules.json
 // file is also updated. The JSON file is used by the oxlint-migrate
 // and eslint-plugin-oxlint repos to keep everything synced.
-const VITEST_COMPATIBLE_JEST_RULES: [&str; 26] = [
+const VITEST_COMPATIBLE_JEST_RULES: [&str; 25] = [
     "no-restricted-jest-methods",
     "no-restricted-matchers",
     "no-standalone-expect",
@@ -63,7 +63,6 @@ const VITEST_COMPATIBLE_JEST_RULES: [&str; 26] = [
     "require-hook",
     "require-to-throw-message",
     "valid-describe-callback",
-    "valid-expect-in-promise",
 ];
 
 /// List of Eslint rules that have TypeScript equivalents.
